moved implicit class test to own dir
